add script

+++ scripts/reset-proj.ps1
@@ -1,8 +1,18 @@
 param(
-  [switch]$lockfile = $false
+  [switch]$lockfile = $false,
+  [switch]$start = $false,
+  [switch]$build = $false
 )
+# echo lockfile: $lockfile start: $start
 Remove-Item .\node_modules\ -Recurse
-Remove-Item .\package-lock.json
+if ($lockfile) {
+  Remove-Item .\package-lock.json
+}
 npm install
 gatsby clean
-
+if ($start) {
+  npm run start
+}
+if ($build) {
+  npm run build
+}
